Rename schedule.{cpp,h} to loopnest.{cpp,h}.
[ghstack-poisoned]
