+++ torch/csrc/jit/tensorexpr/kernel.cpp
@@ -4,7 +4,7 @@
 #include <torch/csrc/jit/tensorexpr/analysis.h>
 #include <torch/csrc/jit/tensorexpr/ir_printer.h>
 #include <torch/csrc/jit/tensorexpr/ir_simplifier.h>
-#include <torch/csrc/jit/tensorexpr/schedule.h>
+#include <torch/csrc/jit/tensorexpr/loopnest.h>
 
 using namespace torch::jit;
 using namespace torch::jit::tensorexpr;
